Extract Store into decorator
