+++ .babelrc
@@ -1,6 +1,9 @@
 {
   "presets": [
     "es2015",
-    "stage-0"
+    "stage-0",
+  ],
+  "plugins": [
+    "transform-decorators-legacy",
   ]
 }
